requirements.txt
newsapi-python==0.2.7
beautifulsoup4==4.9.3
gensim==3.8.3
requests==2.25.1
transformers==4.37.2
langchain==0.0.336
google-generativeai==0.2.2